Move inquiries API path from /exp to /v1
