+++ st2api/tests/base.py
@@ -256,17 +256,17 @@ class BaseInquiryControllerTestCase(BaseFunctionalTest, CleanDbTestCase):
         return self.app.get('/v1/executions/%s' % actionexecution_id, *args, **kwargs)
 
     def _do_get_one(self, inquiry_id, *args, **kwargs):
-        return self.app.get('/exp/inquiries/%s' % inquiry_id, *args, **kwargs)
+        return self.app.get('/v1/inquiries/%s' % inquiry_id, *args, **kwargs)
 
     def _do_get_all(self, limit=50, *args, **kwargs):
-        return self.app.get('/exp/inquiries/?limit=%s' % limit, *args, **kwargs)
+        return self.app.get('/v1/inquiries/?limit=%s' % limit, *args, **kwargs)
 
     def _do_respond(self, inquiry_id, response, *args, **kwargs):
         payload = {
             "id": inquiry_id,
             "response": response
         }
-        return self.app.put_json('/exp/inquiries/%s' % inquiry_id, payload, *args, **kwargs)
+        return self.app.put_json('/v1/inquiries/%s' % inquiry_id, payload, *args, **kwargs)
 
     def _do_create_inquiry(self, liveaction, result, status='pending', *args, **kwargs):
         post_resp = self.app.post_json('/v1/executions', liveaction, *args, **kwargs)
