+++ st2common/st2common/openapi.yaml.j2
@@ -1357,9 +1357,9 @@ paths:
           description: Unexpected error
           schema:
             $ref: '#/definitions/Error'
-  /api/exp/inquiries:
+  /api/v1/inquiries:
     get:
-      operationId: st2api.controllers.exp.inquiries:inquiries_controller.get_all
+      operationId: st2api.controllers.v1.inquiries:inquiries_controller.get_all
       x-permissions: {{ PERMISSION_TYPE.INQUIRY_LIST }}
       description: Returns a list of all inquiries
       parameters:
@@ -1397,9 +1397,9 @@ paths:
           description: Unexpected error
           schema:
             $ref: '#/definitions/Error'
-  /api/exp/inquiries/{inquiry_id}:
+  /api/v1/inquiries/{inquiry_id}:
     get:
-      operationId: st2api.controllers.exp.inquiries:inquiries_controller.get_one
+      operationId: st2api.controllers.v1.inquiries:inquiries_controller.get_one
       description: |
         Get a specific Inquiry based on ID.
       parameters:
@@ -1423,7 +1423,7 @@ paths:
           schema:
             $ref: '#/definitions/Error'
     put:
-      operationId: st2api.controllers.exp.inquiries:inquiries_controller.put
+      operationId: st2api.controllers.v1.inquiries:inquiries_controller.put
       description: |
         Respond to an Inquiry.
       parameters:
